refactor guard and add decorators
