+++ server/src/auth/auth.controller.ts
@@ -4,7 +4,8 @@ import { AuthDto, SignInDto } from './dto';
 import { Tokens } from './types';
 import { AuthGuard } from '@nestjs/passport';
 import { Request } from "express"
-import { AtGuard, RtGuard } from './guards';
+import { AtGuard, RtGuard } from '../common/guards';
+import { GetCurrentUser, GetCurrentUserId } from '../common/decorators';
 
 @Controller('auth')
 export class AuthController {
@@ -25,16 +26,14 @@ export class AuthController {
     @UseGuards(AtGuard)
     @Post('logout')
     @HttpCode(HttpStatus.OK)
-    logout(@Req() req: Request) {
-        const user = req.user;
-        return this.authService.logout(user["sub"])
+    logout(@GetCurrentUserId() userId: string) {
+        return this.authService.logout(userId)
     }
 
     @UseGuards(RtGuard)
     @Post('refresh')
     @HttpCode(HttpStatus.OK)
-    refreshTokens(@Req() req: Request) {
-        const user = req.user;
-        return this.authService.refreshTokens(user["sub"], user["refreshToken"])
+    refreshTokens(@GetCurrentUserId() userId: string, @GetCurrentUser("refreshToken") refreshToken: string) {
+        return this.authService.refreshTokens(userId, refreshToken)
     }
 }
