+++ server/src/auth/types/index.ts
@@ -1,1 +1,3 @@
 export * from './tokens.types'
+export * from './jwtPayload.type'
+export * from './jwtPayloadWithRt.type'
