+++ server/src/auth/types/jwtPayload.type.ts
@@ -1,0 +1,4 @@
+export type JwtPayload = {
+    email: string;
+    sub: string;
+};
