+++ server/src/auth/types/jwtPayloadWithRt.type.ts
@@ -1,0 +1,3 @@
+import { JwtPayload } from '.';
+
+export type JwtPayloadWithRt = JwtPayload & { refreshToken: string };
