+++ server/src/common/decorators/get-current-user-decorator.ts
@@ -1,0 +1,10 @@
+import { createParamDecorator, ExecutionContext } from '@nestjs/common';
+import { JwtPayloadWithRt } from '../../auth/types';
+
+export const GetCurrentUser = createParamDecorator(
+    (data: keyof JwtPayloadWithRt | undefined, context: ExecutionContext) => {
+        const request = context.switchToHttp().getRequest();
+        if (!data) return request.user;
+        return request.user[data];
+    },
+);
