+++ server/src/common/decorators/get-current-user-id.decorator.ts
@@ -1,0 +1,10 @@
+import { createParamDecorator, ExecutionContext } from '@nestjs/common';
+import { JwtPayload } from '../../auth/types';
+
+export const GetCurrentUserId = createParamDecorator(
+    (_: undefined, context: ExecutionContext): string => {
+        const request = context.switchToHttp().getRequest();
+        const user = request.user as JwtPayload;
+        return user.sub;
+    },
+);
