+++ server/src/common/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './get-current-user-decorator'
+export * from './get-current-user-id.decorator'
